requirements.txt
flask>=2.3.0
waitress>=2.1.0
pywin32>=306
Pillow>=10.0.0
PyPDF2>=3.0.0
python-magic-bin>=0.4.14
loguru>=0.7.0
pydantic>=2.0.0